Revert "Feature/ac 49 standard search"
