+++ it/org/corespring/platform/core/services/SubjectQueryServiceTest.scala
@@ -3,29 +3,20 @@ package org.corespring.platform.core.services
 import org.corespring.it.IntegrationSpecification
 import org.corespring.test.helpers.models.SubjectHelper
 import org.specs2.mutable.BeforeAfter
-import play.api.libs.json.Json
 
 class SubjectQueryServiceTest extends IntegrationSpecification {
 
   "Subject Query Service" should {
 
-    def makeQuery(s:String) = Json.obj("searchTerm" -> s).toString()
-
-    "be able to find single items" in new SubjectData("some test subject") {
-      val result = SubjectQueryService.query(makeQuery("some test"))
+    "handle" in new SubjectData("some test subject") {
+      val result = SubjectQueryService.query("some test")
       result.length === 1
     }
 
-    "be able to find multiple items" in new SubjectData("!! 1", "!! 2") {
-      SubjectQueryService.query(makeQuery("!!")).length === 2
-      SubjectQueryService.query(makeQuery("!!!")).length === 0
+    "handle multiple" in new SubjectData("!! 1", "!! 2") {
+      SubjectQueryService.query("!!").length === 2
+      SubjectQueryService.query("!!!").length === 0
     }
-
-    "ignores case" in new SubjectData("UPPERCASE", "lowercase") {
-      SubjectQueryService.query(makeQuery("case")).length === 2
-      SubjectQueryService.query(makeQuery("CASE")).length === 2
-    }
-
   }
 
   class SubjectData(val subjects: String*) extends BeforeAfter {
